requirements.txt
Django==1.7.7
Flask==0.10.1
Jinja2==2.7.3
MarkupSafe==0.23
MySQL-python==1.2.5
Werkzeug==0.10.1
argparse==1.2.1
djangorestframework==3.1.1
gunicorn==19.3.0
heroku==0.1.4
itsdangerous==0.24
mockup==1.8.2
pbr==0.10.0
python-dateutil==1.5
requests==2.6.0
six==1.8.0
stevedore==1.1.0
virtualenv==1.11.6
virtualenv-clone==0.2.5
virtualenvwrapper==4.3.1
whitenoise==1.0.6
wsgiref==0.1.2
django-cors-headers
